+++ CHANGELOG.md
@@ -17,6 +17,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   [GEOD-1130](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1130)
 - Props in the Map Files is now Defined on import instead of using the full path throughout
   [GEOD-1031](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1031)
+- Added Parameters and Dates to the URL for the Detail Page
+  [GEOD-1173](https://opensource.ncsa.illinois.edu/jira/browse/GEOD-1173)
 
 ### Fixed
 - Log error whe sensor is properly defined and we open map popup.